requirements.txt
python-dotenv>=1.0.0
requests>=2.0.0
playsound>=1.3.0
psutil>=5.9.0
pytest>=7.0.0
pytest>=7.0.0
